requirements.txt
python-fasthtml==0.12.1
MonsterUI==1.0.10
pydantic==2.10.6
pydantic-settings==2.8.0
python-dotenv==1.0.1
workalendar==17.0.*
plotly[express]==6.0.*
pandas==2.2.*